+++ contao/dca/tl_user.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * Copyright (c) 2021 Heimrich & Hannot GmbH
  *
@@ -8,21 +10,14 @@
 
 use Contao\CoreBundle\DataContainer\PaletteManipulator;
 
-$dca = &$GLOBALS['TL_DCA']['tl_user_group'];
+$dca = &$GLOBALS['TL_DCA']['tl_user'];
 
 PaletteManipulator::create()
     ->addLegend('huhAdvDash_legend', 'forms_legend', PaletteManipulator::POSITION_AFTER, true)
-//    ->addField('huhAdvDash_visibleUserGroups', 'huhAdvDash_legend', PaletteManipulator::POSITION_APPEND)
     ->addField('huhAdvDash_versionsRights', 'huhAdvDash_legend', PaletteManipulator::POSITION_APPEND)
-    ->applyToPalette('default', 'tl_user_group');
-
-//$dca['fields']['huhAdvDash_visibleUserGroups'] = [
-//    'exclude'    => true,
-//    'inputType'  => 'checkbox',
-//    'foreignKey' => 'tl_user_group.name',
-//    'eval'       => ['multiple' => true],
-//    'sql'        => "blob NULL"
-//];
+    ->applyToPalette('extend', 'tl_user')
+    ->applyToPalette('custom', 'tl_user')
+;
 
 $dca['fields']['huhAdvDash_versionsRights'] = [
     'exclude' => true,