+++ src/DependencyInjection/Configuration.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * Copyright (c) 2021 Heimrich & Hannot GmbH
  *
@@ -8,14 +10,13 @@
 
 namespace HeimrichHannot\AdvancedDashboardBundle\DependencyInjection;
 
-use HeimrichHannot\AdvancedDashboardBundle\VersionList\VersionListConfiguration;
-use HeimrichHannot\AdvancedDashboardBundle\VersionList\VersionListGenerator;
+use HeimrichHannot\AdvancedDashboardBundle\VersionList\AccessLevel;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
 class Configuration implements ConfigurationInterface
 {
-    public function getConfigTreeBuilder()
+    public function getConfigTreeBuilder(): TreeBuilder
     {
         $treeBuilder = new TreeBuilder('huh_advanced_dashboard');
 
@@ -23,23 +24,20 @@ class Configuration implements ConfigurationInterface
             ->children()
                 ->arrayNode('versions_rights')
                     ->info('Configure user rights for version list. Can be selected in the user and user group settings.')
+                    ->defaultValue([])
                     ->useAttributeAsKey('name')
                     ->arrayPrototype()
                         ->info("The title of the configuration. Should be a unique alias/name containing just 'a-z0-9-_' like 'all_users','editor_news'.")
                         ->children()
-                            ->arrayNode('columns')
-                                ->info('Allowed version table columns. Empty means all columns are allowed.')
-                                ->defaultValue(array_keys(VersionListGenerator::columns()))
-                                ->scalarPrototype()->end()
-                            ->end()
                             ->arrayNode('tables')
                                 ->info('Allowed database tables. Empty means all tables are allowed.')
+                                ->defaultValue([])
                                 ->scalarPrototype()->end()
                             ->end()
                             ->enumNode('user_access_level')
                                 ->info('Access rights for other users version logs.')
-                                ->values([VersionListConfiguration::USER_ACCESS_LEVEL_ALL, VersionListConfiguration::USER_ACCESS_LEVEL_SELF])
-                                ->defaultValue([VersionListConfiguration::USER_ACCESS_LEVEL_SELF])
+                                ->values(array_column(AccessLevel::cases(), 'value'))
+                                ->defaultValue(AccessLevel::SELF->value)
                             ->end()
                         ->end()
                     ->end()