+++ src/DependencyInjection/HeimrichHannotAdvancedDashboardExtension.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * Copyright (c) 2021 Heimrich & Hannot GmbH
  *
@@ -8,8 +10,7 @@
 
 namespace HeimrichHannot\AdvancedDashboardBundle\DependencyInjection;
 
-use HeimrichHannot\AdvancedDashboardBundle\VersionList\VersionListConfiguration;
-use HeimrichHannot\AdvancedDashboardBundle\VersionList\VersionListGenerator;
+use HeimrichHannot\AdvancedDashboardBundle\VersionList\AccessLevel;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
@@ -17,28 +18,29 @@ use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
 class HeimrichHannotAdvancedDashboardExtension extends Extension
 {
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $loader = new YamlFileLoader(
             $container,
-            new FileLocator(__DIR__.'/../Resources/config')
+            new FileLocator(__DIR__.'/../../config')
         );
-        $loader->load('services.yml');
+        $loader->load('services.yaml');
 
-        array_unshift($configs, ['versions_rights' => ['default' => []]]);
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
 
-        if (!isset($config['versions_rights']['default'])) {
-            $config['versions_rights']['default'] = [
-                'user_access_level' => VersionListConfiguration::USER_ACCESS_LEVEL_SELF,
-                'columns' => array_keys(VersionListGenerator::columns()),
-            ];
-        }
+        $config['versions_rights']['default'] = array_replace(
+            [
+                'user_access_level' => AccessLevel::SELF->value,
+                'tables' => [],
+            ],
+            $config['versions_rights']['default'] ?? [],
+        );
+
         $container->setParameter('huh_advanced_dashboard', $config);
     }
 
-    public function getAlias()
+    public function getAlias(): string
     {
         return 'huh_advanced_dashboard';
     }