+++ src/VersionList/AccessLevel.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\AdvancedDashboardBundle\VersionList;
+
+enum AccessLevel: string
+{
+    case ALL = 'all';
+    case SELF = 'self';
+}