+++ translations/contao_tl_user_group.de.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+return [
+    'tl_user_group' => [
+        'huhAdvDash_legend' => 'Erweitertes Dashboard - Rechte',
+        'huhAdvDash_versionsRights' => ['Versions-Rechte', 'Wählen Sie hier die Versions-Rechte aus.'],
+    ],
+];